Refactor project index coordinator into dedicated module
