+++ src/plugin/src/project-index/index.js
@@ -46,6 +46,7 @@ import {
     createFileScopeDescriptor
 } from "./resource-analysis.js";
 import { loadBuiltInIdentifiers } from "./built-in-identifiers.js";
+import { createProjectIndexCoordinatorFactory } from "./coordinator.js";
 
 const defaultProjectIndexParser = getDefaultProjectIndexParser();
 
@@ -102,118 +103,6 @@ function resolveProjectIndexParser(options) {
     );
 }
 
-function normalizeEnsureReadyDescriptor(descriptor) {
-    const projectRoot = descriptor?.projectRoot;
-    if (!projectRoot) {
-        throw new Error("projectRoot must be provided to ensureReady");
-    }
-
-    return {
-        descriptor,
-        resolvedRoot: path.resolve(projectRoot)
-    };
-}
-
-function resolveEnsureReadyContext({
-    descriptor,
-    abortController,
-    disposedMessage
-}) {
-    const { resolvedRoot } = normalizeEnsureReadyDescriptor(descriptor);
-    const signal = abortController.signal;
-    throwIfAborted(signal, disposedMessage);
-
-    return {
-        descriptor,
-        resolvedRoot,
-        key: resolvedRoot,
-        signal
-    };
-}
-
-function trackInFlightOperation(map, key, createOperation) {
-    if (map.has(key)) {
-        return map.get(key);
-    }
-
-    const pending = Promise.resolve()
-        .then(createOperation)
-        .finally(() => {
-            map.delete(key);
-        });
-
-    map.set(key, pending);
-    return pending;
-}
-
-async function executeEnsureReadyOperation({
-    descriptor,
-    resolvedRoot,
-    signal,
-    fsFacade,
-    loadCache,
-    saveCache,
-    buildIndex,
-    cacheMaxSizeBytes,
-    disposedMessage
-}) {
-    const descriptorOptions = descriptor ?? {};
-    const loadResult = await loadCache(
-        { ...descriptorOptions, projectRoot: resolvedRoot },
-        fsFacade,
-        { signal }
-    );
-    throwIfAborted(signal, disposedMessage);
-
-    if (loadResult.status === "hit") {
-        throwIfAborted(signal, disposedMessage);
-        return {
-            source: "cache",
-            projectIndex: loadResult.projectIndex,
-            cache: loadResult
-        };
-    }
-
-    const projectIndex = await buildIndex(resolvedRoot, fsFacade, {
-        ...descriptorOptions?.buildOptions,
-        signal
-    });
-    throwIfAborted(signal, disposedMessage);
-
-    const descriptorMaxSizeBytes =
-        descriptorOptions?.maxSizeBytes === undefined
-            ? cacheMaxSizeBytes
-            : descriptorOptions.maxSizeBytes;
-
-    const saveResult = await saveCache(
-        {
-            ...descriptorOptions,
-            projectRoot: resolvedRoot,
-            projectIndex,
-            metricsSummary: projectIndex.metrics,
-            maxSizeBytes: descriptorMaxSizeBytes
-        },
-        fsFacade,
-        { signal }
-    ).catch((error) => {
-        return {
-            status: "failed",
-            error,
-            cacheFilePath: loadResult.cacheFilePath
-        };
-    });
-    throwIfAborted(signal, disposedMessage);
-
-    return {
-        source: "build",
-        projectIndex,
-        cache: {
-            ...loadResult,
-            saveResult
-        }
-    };
-}
-
 export async function findProjectRoot(options, fsFacade = defaultFsFacade) {
     const filepath = options?.filepath;
     const { signal, ensureNotAborted } = createAbortGuard(options, {
@@ -240,76 +129,16 @@ export async function findProjectRoot(options, fsFacade = defaultFsFacade) {
     return null;
 }
 
-export function createProjectIndexCoordinator(options = {}) {
-    const {
-        fsFacade = defaultFsFacade,
-        loadCache = loadProjectIndexCache,
-        saveCache = saveProjectIndexCache,
-        buildIndex = buildProjectIndex,
-        cacheMaxSizeBytes: rawCacheMaxSizeBytes
-    } = options;
+export const createProjectIndexCoordinator =
+    createProjectIndexCoordinatorFactory({
+        defaultFsFacade,
+        defaultLoadCache: loadProjectIndexCache,
+        defaultSaveCache: saveProjectIndexCache,
+        defaultBuildIndex: buildProjectIndex,
+        getDefaultCacheMaxSize: getDefaultProjectIndexCacheMaxSize
+    });
 
-    const cacheMaxSizeBytes =
-        rawCacheMaxSizeBytes === undefined
-            ? getDefaultProjectIndexCacheMaxSize()
-            : rawCacheMaxSizeBytes;
-
-    const inFlight = new Map();
-    let disposed = false;
-    const abortController = new AbortController();
-    const DISPOSED_MESSAGE = "ProjectIndexCoordinator has been disposed";
-
-    function createDisposedError() {
-        return new Error(DISPOSED_MESSAGE);
-    }
-
-    function ensureNotDisposed() {
-        if (disposed) {
-            throw createDisposedError();
-        }
-        throwIfAborted(abortController.signal, DISPOSED_MESSAGE);
-    }
-
-    async function ensureReady(descriptor) {
-        ensureNotDisposed();
-        const context = resolveEnsureReadyContext({
-            descriptor,
-            abortController,
-            disposedMessage: DISPOSED_MESSAGE
-        });
-
-        return trackInFlightOperation(inFlight, context.key, () =>
-            executeEnsureReadyOperation({
-                descriptor,
-                resolvedRoot: context.resolvedRoot,
-                signal: context.signal,
-                fsFacade,
-                loadCache,
-                saveCache,
-                buildIndex,
-                cacheMaxSizeBytes,
-                disposedMessage: DISPOSED_MESSAGE
-            })
-        );
-    }
-
-    function dispose() {
-        if (disposed) {
-            return;
-        }
-
-        disposed = true;
-        if (!abortController.signal.aborted) {
-            abortController.abort(createDisposedError());
-        }
-        inFlight.clear();
-    }
-
-    return {
-        ensureReady,
-        dispose
-    };
-}
+export { createProjectIndexCoordinatorFactory } from "./coordinator.js";
 
 export {
     PROJECT_MANIFEST_EXTENSION,
